+++ test/controllers/membership_controller_test.exs
@@ -90,6 +90,18 @@ defmodule Ask.MembershipControllerTest do
     assert updated_membership.level == "reader"
   end
 
+  test "does not change level to an invalid value", %{conn: conn, user: user} do
+    project = create_project_for_user(user)
+    collaborator_email = "user2@surveda.instedd.org"
+    collaborator = insert(:user, name: "user2", email: collaborator_email)
+    collaborator_membership = %{"user_id" => collaborator.id, "project_id" => project.id, "level" => "editor"}
+    ProjectMembership.changeset(%ProjectMembership{}, collaborator_membership) |> Repo.insert
+
+    put conn, project_membership_update_path(conn, :update, project.id), email: collaborator_email, level: "invalid"
+    updated_membership = Repo.one(from p in ProjectMembership, where: p.user_id == ^collaborator.id)
+    assert updated_membership.level == "editor"
+  end
+
   test "forbids user to update owner", %{conn: conn, user: user} do
     project = create_project_for_user(user)
     collaborator_email = "user2@surveda.instedd.org"
@@ -99,6 +111,46 @@ defmodule Ask.MembershipControllerTest do
 
     assert_error_sent :forbidden, fn ->
       put conn, project_membership_update_path(conn, :update, project.id), email: collaborator_email, level: "reader"
+    end
+  end
+
+  test "forbids editor to upgrade other to owner", %{conn: conn, user: user} do
+    collaborator_email = "user2@surveda.instedd.org"
+    collaborator = insert(:user, name: "user2", email: collaborator_email)
+
+    project = create_project_for_user(collaborator)
+
+    user_membership = %{"user_id" => user.id, "project_id" => project.id, "level" => "editor"}
+    ProjectMembership.changeset(%ProjectMembership{}, user_membership) |> Repo.insert
+
+    assert_error_sent :forbidden, fn ->
+      put conn, project_membership_update_path(conn, :update, project.id), email: user.email, level: "owner"
+    end
+  end
+
+  test "forbids reader to upgrade other to editor", %{conn: conn, user: user} do
+    collaborator_email = "user2@surveda.instedd.org"
+    collaborator = insert(:user, name: "user2", email: collaborator_email)
+
+    project = create_project_for_user(collaborator)
+
+    user_membership = %{"user_id" => user.id, "project_id" => project.id, "level" => "reader"}
+    ProjectMembership.changeset(%ProjectMembership{}, user_membership) |> Repo.insert
+
+    assert_error_sent :forbidden, fn ->
+      put conn, project_membership_update_path(conn, :update, project.id), email: user.email, level: "editor"
+    end
+  end
+
+  test "forbids editor to update to admin", %{conn: conn, user: user} do
+    project = create_project_for_user(user, level: "editor")
+    collaborator_email = "user2@surveda.instedd.org"
+    collaborator = insert(:user, name: "user2", email: collaborator_email)
+    collaborator_membership = %{"user_id" => collaborator.id, "project_id" => project.id, "level" => "admin"}
+    ProjectMembership.changeset(%ProjectMembership{}, collaborator_membership) |> Repo.insert
+
+    assert_error_sent :forbidden, fn ->
+      put conn, project_membership_update_path(conn, :update, project.id), email: collaborator_email, level: "admin"
     end
   end